Move server/server.js to index.js; simplify
